Handle account client and user info states when opening main activity
